Split GraphSuite; simplify LocalSparkContext
